+++ next_action/cli.py
@@ -23,5 +23,7 @@ def next_action() -> None:
             return
         with todotxt_file:
             tasks.extend([Task(line.strip()) for line in todotxt_file.readlines() if line.strip()])
-    actions = next_actions(tasks, set(arguments.contexts), set(arguments.projects), set(arguments.excluded_contexts))
+    actions = next_actions(tasks,
+                           set(arguments.contexts), set(arguments.projects),
+                           set(arguments.excluded_contexts), set(arguments.excluded_projects))
     print("\n".join(action.text for action in actions[:arguments.number]) if actions else "Nothing to do!")